feat: add new skeleton handlers for bff and refactor get-splash to follow file naming convention
