+++ backend-for-frontend/handlers/game-start.js
@@ -1,0 +1,7 @@
+export const getGameStart = () => {
+
+}
+
+export const postGameStart = () => {
+
+}
