+++ backend-for-frontend/handlers/new-game.js
@@ -1,0 +1,7 @@
+export const getNewGame = () => {
+
+}
+
+export const postNewGame = () => {
+
+}
