+++ backend-for-frontend/handlers/player-join.js
@@ -1,0 +1,7 @@
+export const getPlayerJoin = () => {
+  
+}
+
+export const postPlayerJoin = () => {
+  
+}
