+++ backend-for-frontend/handlers/splash.js
